Add the project into a Docker
