+++ API/models/file.py
@@ -1,0 +1,2 @@
+import ultralytics
+print(ultralytics.__version__)
