+++ application/Dockerfile
@@ -1,0 +1,14 @@
+FROM node:20-alpine
+
+WORKDIR /app
+
+COPY package*.json ./
+RUN npm install
+
+COPY . .
+
+RUN chmod +x /app/entrypoint.sh
+
+EXPOSE 3000
+
+ENTRYPOINT ["/app/entrypoint.sh"]
