+++ application/src/app/api/assinaturas/route.js
@@ -1,4 +1,4 @@
-import { Assinatura } from "@/app/database/models/tables";
+import { Assinatura } from "../../../../database/models/tables";
 import { NextResponse } from "next/server";
 
 // Create
