+++ application/src/app/api/tentativaPuzzles/route.js
@@ -1,4 +1,4 @@
-import { TentativaPuzzle } from "@/app/database/models/tables";
+import { TentativaPuzzle } from "../../../../database/models/tables";
 import { NextResponse } from "next/server";
 
 // Read all
